rename files to semiOrderedLike...
