+++ src/main/kotlin/com/tegonal/variist/generators/impl/internalSemiOrderedLikeTransformations.kt
