+++ src/main/kotlin/com/tegonal/variist/generators/impl/semiOrderedLikeArgsGeneratorConcatenators.kt
@@ -1,4 +1,3 @@
-//TODO 3.0.0 rename file to semiOrderedLikeArgsGeneratorConcatenators
 package com.tegonal.variist.generators.impl
 
 import com.tegonal.variist.config._components
